+++ source/extensions/omni.isaac.benchmark_environments/omni/isaac/benchmark_environments/object_interface.py
@@ -14,24 +14,13 @@ import asyncio
 import uuid
 import omni.kit.app
 from omni.isaac.core.utils.prims import delete_prim
+from omni.isaac.core.utils.rotations import matrix_to_euler_angles, euler_angles_to_quat
+from omni.isaac.core.objects import cuboid, sphere, capsule
+from omni.isaac.core.prims.xform_prim import XFormPrim
 
 
-def get_Gf_transform(translation, rotation):
-    """
-    This function converts from np arrays to a Gf.Matrix4d object.  
-    """
-
-    rot_mat = Gf.Matrix3d()
-    for i in range(3):
-        rot_mat.SetColumn(i, Gf.Vec3d(*rotation[i].astype(np.float64)))
-
-    trans = Gf.Vec3d(*translation.astype(np.float64))
-
-    mat = Gf.Matrix4d()
-    mat.SetTranslate(trans)
-    mat.SetRotateOnly(rot_mat)
-
-    return mat
+def matrix_to_quat(rot_mat):
+    return euler_angles_to_quat(matrix_to_euler_angles(rot_mat))
 
 
 class Object:
@@ -55,12 +44,10 @@ class Object:
     TODO: allow meshes to be loaded for an object
     """
 
-    def __init__(self, _stage, base_translation, base_rotation, color=Gf.Vec3f(1.0, 1.0, 0), **kwargs):
-        self._stage = _stage
+    def __init__(self, base_translation, base_rotation, color=np.array([1.0, 1.0, 0]), **kwargs):
 
         self.initial_base_trans = base_translation
         self.initial_base_rotation = base_rotation
-        self.base_pose = get_Gf_transform(base_translation, base_rotation)
         self.components = []
         self.color = color
 
@@ -69,145 +56,91 @@ class Object:
 
         # make a random USD path for all the prims in this object
         object_id = str(uuid.uuid4())
-        self.base_path = ("/scene/object-" + object_id + "-").replace("-", "_")
-        self.target_path = ("/scene/target-" + object_id + "-").replace("-", "_")
+        self.base_path = ("/scene/object-" + object_id).replace("-", "_")
+        self.base_xform = XFormPrim(self.base_path)
+        self.base_xform.set_world_pose(base_translation, matrix_to_quat(base_rotation))
 
         self.prim2pose_rel = {}  # pose of each component relative to the base
 
         self.construct(**kwargs)
+        self.set_physics_properties(enable_rigid_body=False, enable_collisions=True)
 
     def construct(self, **kwargs):
         pass
 
     def get_random_target(self, make_visible=True):
         if self.last_target is not None:
-            self.last_target.MakeInvisible()
+            self.last_target.set_visibility(False)
 
         if len(self.targets) > 0:
             target = np.random.choice(self.targets)
-            if make_visible:
-                target.MakeVisible()
+            target.set_visibility(make_visible)
             self.last_target = target
             return target
 
-    def get_all_prims(self):
+    def get_all_components(self):
         # get all prims that the robot should avoid
-        return [com.GetPrim() for com in self.components]
+        return self.components
 
     def create_target(
         self,
         relative_translation=np.zeros(3),
-        relative_rotation=None,
-        target_color=Gf.Vec3f(1.0, 0.0, 0.0),
+        relative_rotation=np.eye(3),
+        target_color=np.array([1.0, 0, 0]),
         target_size=5,
     ):
 
-        path = self.target_path + str(len(self.targets))
-        geom = UsdGeom.Cube.Define(self._stage, path)
+        path = self.base_path + "/target_" + str(len(self.targets))
+        target = cuboid.VisualCuboid(path, size=target_size * np.ones(3), color=target_color)
 
-        if relative_rotation is None:
-            rel_pose = Gf.Vec4d(*relative_translation.astype(np.float64), 1)
-            abs_pos = Gf.Vec3d(*list(rel_pose * self.base_pose)[:3])
-            geom.AddTranslateOp().Set(abs_pos)
-        else:
-            rel_pose = get_Gf_transform(relative_translation, relative_rotation)
-            geom.AddTransformOp().Set(rel_pose * self.base_pose)
-        self.prim2pose_rel[geom] = rel_pose
+        # self.prim2pose_rel[target] = rel_pose
+        target.set_local_pose(relative_translation, matrix_to_quat(relative_rotation))
 
-        geom.CreateSizeAttr(target_size)
+        target.set_visibility(False)
 
-        geom.CreateDisplayColorAttr().Set([target_color])
-        geom.GetPrim().SetCustomDataByKey("type", "target")
+        self.targets.append(target)
 
-        geom.MakeInvisible()
+        return target
 
-        self.targets.append(geom)
+    def create_block(self, size, relative_translation=np.zeros(3), relative_rotation=np.eye(3)):
+        path = self.base_path + "/cuboid_" + str(len(self.components))
+        cube = cuboid.DynamicCuboid(path, size=size, color=self.color)
+        cube.set_local_pose(relative_translation, matrix_to_quat(relative_rotation))
 
-        return geom
+        self.components.append(cube)
 
-    def create_block(self, size, scales, relative_translation=np.zeros(3), relative_rotation=np.eye(3)):
-        path = self.base_path + str(len(self.components))
-        geom = UsdGeom.Cube.Define(self._stage, path)
-
-        rel_pose = get_Gf_transform(relative_translation, relative_rotation)
-        self.prim2pose_rel[geom] = rel_pose
-
-        geom.AddTransformOp().Set(rel_pose * self.base_pose)
-
-        geom.CreateSizeAttr(size)
-        scales = Gf.Vec3f(scales[0], scales[1], scales[2])
-        geom.AddScaleOp().Set(scales)
-
-        geom.CreateDisplayColorAttr().Set([self.color])
-        geom.GetPrim().SetCustomDataByKey("type", "box")
-
-        self.components.append(geom)
-
-        return geom
+        return cube
 
     def create_sphere(self, radius, relative_translation=np.zeros(3), relative_rotation=np.eye(3)):
-        path = self.base_path + str(len(self.components))
-        geom = UsdGeom.Sphere.Define(self._stage, path)
-        geom.CreateRadiusAttr(radius)
+        path = self.base_path + "/sphere_" + str(len(self.components))
+        sphere = cuboid.DynamicSphere(path, radius=radius, color=self.color)
+        sphere.set_local_pose(relative_translation, matrix_to_quat(relative_rotation))
 
-        geom.CreateDisplayColorAttr().Set([self.color])
+        self.components.append(sphere)
 
-        rel_pose = get_Gf_transform(relative_translation, relative_rotation)
-
-        self.prim2pose_rel[geom] = rel_pose
-
-        geom.AddTransformOp().Set(rel_pose * self.base_pose)
-        geom.GetPrim().SetCustomDataByKey("type", "sphere")
-
-        self.components.append(geom)
-
-        return geom
+        return sphere
 
     def create_capsule(self, radius, height, relative_translation=np.zeros(3), relative_rotation=np.eye(3)):
-        path = self.base_path + str(len(self.components))
-        geom = UsdGeom.Capsule.Define(self._stage, path)
-        geom.CreateRadiusAttr(radius)
-        geom.CreateHeightAttr(height)
+        path = self.base_path + "/capsule_" + str(len(self.components))
+        capsule = cuboid.DynamicSphere(path, radius=radius, height=height, color=self.color)
+        capsule.set_local_pose(relative_translation, matrix_to_quat(relative_rotation))
 
-        geom.CreateDisplayColorAttr().Set([self.color])
+        self.components.append(capsule)
 
-        rel_pose = get_Gf_transform(relative_translation, relative_rotation)
-
-        self.prim2pose_rel[geom] = rel_pose
-
-        geom.AddTransformOp().Set(rel_pose * self.base_pose)
-        geom.GetPrim().SetCustomDataByKey("type", "capsule")
-
-        self.components.append(geom)
-
-        return geom
+        return capsule
 
     def set_base_pose(self, translation=np.zeros(3), rotation=np.eye(3)):
-        self.base_pose = get_Gf_transform(translation, rotation)
 
-        for component in self.components:
-            component.GetPrim().GetAttribute("xformOp:transform").Set(self.prim2pose_rel[component] * self.base_pose)
+        self.base_xform.set_world_pose(translation, matrix_to_quat(rotation))
 
-        for target in self.targets:
-            target_prim = target.GetPrim()
-            if target_prim.HasAttribute("xformOp:transform"):
-                target_prim.GetAttribute("xformOp:transform").Set(self.prim2pose_rel[target] * self.base_pose)
-            else:
-                rel_pos = self.prim2pose_rel[target]
-                abs_pos = Gf.Vec3d(*list(rel_pos * self.base_pose)[:3])
-                target_prim.GetAttribute("xformOp:translate").Set(abs_pos)
-
-    def set_color(self, color):
-        self.color = color
-        for component in self.components:
-            components.GetDisplayColorAttr().Set([self.color])
+    # def set_color(self, color):
+    #     self.color = color
+    #     for component in self.components:
+    #         component.GetDisplayColorAttr().Set([self.color])
 
     def set_visibility(self, on=True):
         for component in self.components:
-            if on:
-                component.MakeVisible()
-            else:
-                component.MakeInvisible()
+            component.set_visibility(on)
 
     def set_physics_properties(self, enable_rigid_body=True, enable_collisions=True, mass=1e-7):
         """
@@ -224,33 +157,29 @@ class Object:
 
             Enabling only collisions will make this Object become impassable for the robot
         """
-
-        # async def set_physics_properties_async():
-        # await omni.kit.app.get_app().next_update_async()
         for component in self.components:
             if enable_rigid_body:
-                UsdPhysics.RigidBodyAPI.Apply(component.GetPrim())
-                # await omni.kit.app.get_app().next_update_async()
-            if enable_collisions:
-                UsdPhysics.CollisionAPI.Apply(component.GetPrim())
-                # await omni.kit.app.get_app().next_update_async()
-            massAPI = UsdPhysics.MassAPI.Apply(component.GetPrim())
-            massAPI.CreateMassAttr(mass)
-            # await omni.kit.app.get_app().next_update_async()
+                component.enable_rigid_body_physics()
+            else:
+                component.disable_rigid_body_physics()
 
-        # asyncio.ensure_future(set_physics_properties_async())
+            if enable_collisions:
+                # right now cannot be changed due to pending changes to core
+                pass
+
+            component.set_mass(mass)
 
     def delete(self):
         for component in self.components:
-            delete_prim(component.GetPath())
+            delete_prim(component.prim_path)
         self.components = []
 
         for target in self.targets:
-            delete_prim(target.GetPath())
+            delete_prim(target.prim_path)
 
         self.targets = []
 
     def reset(self):
         self.set_base_pose(self.initial_base_trans, self.initial_base_rotation)
         for target in self.targets:
-            target.MakeInvisible()
+            target.set_visibility(False)