+++ source/extensions/omni.isaac.motion_generation/omni/isaac/motion_generation/motion_policy_interface.py
@@ -9,8 +9,10 @@
 from pxr import Usd, UsdGeom
 from enum import Enum
 import numpy as np
-import copy
+import carb
 import typing
+import omni.isaac.core.objects
+from omni.isaac.core.objects import cuboid, sphere, capsule, cylinder, cone, ground_plane
 
 
 class PolicyType(Enum):
@@ -22,11 +24,7 @@ class MotionPolicy:
     """Interface for implementing motion policies for compatibility with `MotionGenerator` interface.
     """
 
-    def __init__(self, _stage: Usd.Stage, policy_type: PolicyType) -> None:
-        self._stage = _stage
-        self._meters_per_unit = UsdGeom.GetStageMetersPerUnit(self._stage)
-        self._target_prim = None
-        self._target_prim_is_position_only = None
+    def __init__(self, policy_type: PolicyType) -> None:
         self.initialized = False
         self.policy_type = policy_type
 
@@ -112,22 +110,17 @@ class MotionPolicy:
         """
         pass
 
-    def set_end_effector_target(self, target_prim: Usd.Prim, position_only: bool = False) -> None:
+    def set_end_effector_target(self, target_translation=None, target_orientation=None) -> None:
         """Set end effector target.
 
         Args:
-            target_prim (Usd.Prim): USD prim of the target. target_prim may also be None, in which case it is up 
-                to the policy to specify the desired behavior of the robot. Some policies store a default  c-space 
-                configuration in their config files and drive the robot to that position when there is no target 
-                specified.
-            position_only (bool, optional):  When True, the policy will use only the position (not orientation) of the 
-                target_prim as the target. Defaults to False.
+            target_translation (nd.array): Translation vector (3x1) for robot end effector
+            target_orientation (nd.array): Quaternion of desired rotation for robot end effector 
 
         Returns:
             None
         """
-        self._target_prim = target_prim
-        self._target_prim_is_position_only = position_only
+        pass
 
     def get_end_effector_pose(self, joint_positions: np.array) -> typing.Tuple[np.array, np.array, np.array]:
         """Return pose of the end effector at the given end effector position.
@@ -140,175 +133,119 @@ class MotionPolicy:
         """
         pass
 
-    def get_prim_pose(
-        self, prim: Usd.Prim, default_trans: np.array = np.zeros(3), default_rot: np.array = np.eye(3)
-    ) -> bool:
-        """Return pose of prim.
-        
-        USD prims that lack translational information are placed on the stage at the point (0,0,0). USD prims that lack 
-        rotational information are placed on the stage with the identity rotation. Reading translation/rotation 
-        information from prims directly yields a 4x4 transform matrix, which necessitates filling in missing information 
-        with default information. This method allows the caller to know what information is actually present in the USD 
-        prims by passing in None for the defaults.
+    def add_cuboid(self, cuboid: cuboid.DynamicCuboid, static: bool = False) -> bool:
+        """Add a block obstacle.
 
         Args:
-            prim (Usd.Prim): Prim for which pose will be returned.
-            default_trans (np.array, optional): Translation component of pose to be returned
-                if the prim has no translational information. Defaults to np.zeros(3).
-            default_rot (cp.array, optional): Rotational component of pose to be returned
-                if the prim contains no rotational information. Defaults to np.eye(3).
-
-        Returns:
-            Tuple[np.array, np.array]: Prim pose returned as translation vector (3 x 1) and rotation matrix  (3 x 3).
-        """
-        xform = UsdGeom.Xformable(prim)
-        if xform.GetXformOpOrderAttr().Get() is None:
-            return default_trans, default_rot
-
-        attr_name_base = "xformOp:"
-
-        rotation_attrs = [
-            "transform",
-            "orient",
-            "rotateX",
-            "rotateXYZ",
-            "rotateXZY",
-            "rotateY",
-            "rotateYXZ",
-            "rotateYZX",
-            "rotateZ",
-            "rotateZYX",
-            "rotateZXY",
-        ]
-        rotation_attrs = [attr_name_base + r for r in rotation_attrs]
-
-        translation_attrs = ["transform", "translate"]
-        translation_attrs = [attr_name_base + t for t in translation_attrs]
-
-        has_rot_attr = False
-        has_trans_attr = False
-        for op in xform.GetXformOpOrderAttr().Get():
-            if op in rotation_attrs:
-                has_rot_attr = True
-            if op in translation_attrs:
-                has_trans_attr = True
-
-        mat = xform.GetLocalTransformation()
-
-        if has_trans_attr:
-            trans = np.array(mat.ExtractTranslation())
-        else:
-            trans = default_trans
-
-        # GfMatrix4d() objects have transposed rotation matrices
-        # rotation matrix may not have normalized rows due to xformOp:scale
-        if has_rot_attr:
-            rot = np.array(mat.ExtractRotationMatrix()).T
-            row_norms = np.linalg.norm(rot, axis=1)
-            rot = rot / row_norms[:, np.newaxis]
-        else:
-            rot = default_rot
-
-        return self._meters_per_unit * trans, rot
-
-    def create_cube(self, block_prim: Usd.Prim, side_length: float = None, static: bool = False) -> bool:
-        """Create a cube obstacle.
-
-        Args:
-            block_prim (Usd.Prim): USD prim representing the cube. Must have pose information.
-            side_length (float, optional): [description]. Length of each side of the cube. If not specified, 
-                side_length is read from 'size' attribute of block_prim. Defaults to None.
-            static (bool, optional): If True, indicate that cube will never change pose, and may be ignored in internal 
+            cuboid (cuboid.DynamicCuboid): Wrapper object for handling rectangular prism Usd Prims.
+            static (bool, optional): If True, indicate that cuboid will never change pose, and may be ignored in internal 
                 world updates. Defaults to False.
 
         Returns:
-            bool: Return True if block_prim exists and has required attributes (i.e., size must be specified \
-                somewhere, either in block_prim or via side_length param).
+            bool: Return True if underlying MotionPolicy has implemented create_cuboid()
         """
+        carb.log_warning("Function create_cuboid() has not been implemented for this MotionPolicy")
         return False
 
-    def create_block(self, block_prim: Usd.Prim, dimensions: np.array = None, static: bool = False) -> bool:
-        """Create a block obstacle.
+    def add_sphere(self, sphere: sphere.DynamicSphere, static: bool = False) -> bool:
+        """Add a sphere obstacle.
 
-        Args:
-            block_prim (Usd.Prim): USD prim representing the block. Must have pose information.
-            dimensions (np.array, optional): Length of block in (x,y,z) dimensions. If not specified, prim must have 
-                'xformOp:scale' and "size" attribute. Defaults to None.
-            static (bool, optional): If True, indicate that cube will never change pose, and may be ignored in internal 
+        Args:omni.isaac.core.objectshange pose, and may be ignored in internal 
                 world updates. Defaults to False.
 
         Returns:
-            bool: Return True if block_prim exists and has required attributes (i.e., side lengths must be specified \
-                    somewhere, either in block_prim or via dimensions param).
+            bool: Return True if underlying MotionPolicy has implemented create_sphere()
         """
+        carb.log_warning("Function create_sphere() has not been implemented for this MotionPolicy")
         return False
 
-    def create_sphere(self, sphere_prim: Usd.Prim, radius: float = None, static: bool = False) -> bool:
-        """Create a sphere obstacle.
+    def add_capsule(self, capsule: capsule.DynamicCapsule, static: bool = False) -> bool:
+        """Add a capsule obstacle.
 
         Args:
-            sphere_prim (Usd.Prim): USD prim representing the sphere. Must have pose information.
-            radius (float, optional): Radius of the sphere. If not specified, radius is read from 'radius' attribute of 
-                sphere_prim. Defaults to None.
-            static (bool, optional): If True, indicate that cube will never change pose, and may be ignored in internal 
+            capsule (capsule.DynamicCapsule): Wrapper object for handling capsule Usd Prims.
+            static (bool, optional): If True, indicate that capsule will never change pose, and may be ignored in internal 
                 world updates. Defaults to False.
 
         Returns:
-            bool: Return True if sphere_prim exists and has required attributes (i.e., radius must be specified \
-                somewhere, either in sphere_prim or via radius param).
+            bool: Return True if underlying MotionPolicy has implemented create_capsule()
         """
+        carb.log_warning("Function create_capsule() has not been implemented for this MotionPolicy")
         return False
 
-    def create_capsule(
-        self, capsule_prim: Usd.Prim, radius: float = None, height: float = None, static: bool = False
-    ) -> bool:
-        """Create a capsule obstacle.
+    def add_cylinder(self, cylinder: cylinder.DynamicCylinder, static: bool = False) -> bool:
+        """Add a cylinder obstacle.
 
         Args:
-            capsule_prim (Usd.Prim): USD prim representing the capsule. Must have pose information.
-            radius (float, optional): Radius of the capsule. If not specified, radius is read from 'radius' attribute 
-                of  capsule_prim. Defaults to None.
-            height (float, optional): Height of the capsule. If not specified, height is read from 'height' attribute of 
-                capsule_prim. Defaults to None.
-            static (bool, optional): If True, indicate that cube will never change pose, and may be ignored in internal 
+            cylinder (cylinder.DynamicCylinder): Wrapper object for handling cylinder Usd Prims.
+            static (bool, optional): If True, indicate that cylinder will never change pose, and may be ignored in internal 
                 world updates. Defaults to False.
 
         Returns:
-            bool: Return True if capsule_prim exists and has required attributes (i.e., radius and height must be \
-                specified  somewhere, either in capsule_prim or via radius and height params).
+            bool: Return True if underlying MotionPolicy has implemented create_cylinder()
         """
+        carb.log_warning("Function create_cylinder() has not been implemented for this MotionPolicy")
         return False
 
-    def disable_obstacle(self, obstacle_prim: Usd.Prim) -> bool:
+    def add_cone(self, cone: cone.DynamicCone, static: bool = False) -> bool:
+        """Add a cone obstacle.
+
+        Args:
+            cone (cone.DynamicCylinder): Wrapper object for handling cone Usd Prims.
+            static (bool, optional): If True, indicate that cone will never change pose, and may be ignored in internal 
+                world updates. Defaults to False.
+
+        Returns:
+            bool: Return True if underlying MotionPolicy has implemented create_cone()
+        """
+        carb.log_warning("Function create_cone() has not been implemented for this MotionPolicy")
+        return False
+
+    def add_ground_plane(self, ground_plane: ground_plane.GroundPlane) -> bool:
+        """Add a ground_plane
+
+        Args:
+            ground_plane (ground_plane.DynamicCylinder): Wrapper object for handling ground_plane Usd Prims.
+
+        Returns:
+            bool: Return True if underlying MotionPolicy has implemented create_ground_plane()
+        """
+        carb.log_warning("Function create_ground_plane() has not been implemented for this MotionPolicy")
+        return False
+
+    def disable_obstacle(self, obstacle: omni.isaac.core.objects) -> bool:
         """Disable collision avoidance for obstacle.
 
         Args:
-            obstacle_prim (Usd.Prim): USD prim for obstacle to be disabled.
+            obstacle (core.object): obstacle to be disabled.
 
         Returns:
             bool: Return true if obstacle was identified and successfully disabled.
         """
+        carb.log_warning("Function disable_obstacle() has not been implemented for this MotionPolicy")
         return False
 
-    def enable_obstacle(self, obstacle_prim: Usd.Prim) -> bool:
+    def enable_obstacle(self, obstacle: omni.isaac.core.objects) -> bool:
         """Enable collision avoidance for obstacle.
 
         Args:
-            obstacle_prim (Usd.Prim): USD prim for obstacle to be enabled.
+            obstacle (core.object): obstacle to be enabled.
 
         Returns:
             bool: Return true if obstacle was identified and successfully enabled.
         """
+        carb.log_warning("Function enable_obstacle() has not been implemented for this MotionPolicy")
         return False
 
-    def remove_obstacle(self, obstacle_prim: Usd.Prim) -> bool:
+    def remove_obstacle(self, obstacle: omni.isaac.core.objects) -> bool:
         """Remove obstacle from collision avoidance. Obstacle cannot be re-enabled via enable_obstacle() after 
         removal.
         
         Args:
-            obstacle_prim (Usd.Prim): USD prim for obstacle to be removed.
+            obstacle (core.object): obstacle to be removed.
 
         Returns:
             bool: Return true if obstacle was identified and successfully removed.
         """
+        carb.log_warning("Function remove_obstacle() has not been implemented for this MotionPolicy")
         return False